# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  content.CreateNovelRequest:
    properties:
      narration_type:
        description: 旁白类型：narration（旁白/解说）或 dialogue（真人对话）
        type: string
      resource_id:
        description: 资源ID（必填）
        type: string
      style:
        description: 风格：anime（漫剧）、live（真人剧）、mixed（混合）
        type: string
      user_id:
        description: 用户ID（必填）
        type: string
    required:
    - narration_type
    - resource_id
    - style
    - user_id
    type: object
...
  content.GeneratePropImagesRequest:
    properties:
      novel_id:
        description: 小说ID（必填）
        type: string
    required:
    - novel_id
    type: object
...
  http.ErrorResponse:
    properties:
      code:
        description: 错误码（非0表示错误）
        type: integer
      detail:
        description: 错误详情（可选）
        type: string
      message:
        description: 错误消息
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@lemon.ai
    name: API Support
  description: |-
    Lemon is an AI-powered API service built with Eino framework.
    It provides LLM chat, intelligent agent, and more AI capabilities.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Lemon API
  version: "1.0"
paths:
...
  /api/v1/novels:
    get:
      consumes:
      - application/json
      description: 根据当前用户获取剧本列表（分页）
      parameters:
      - description: 页码，默认1
        in: query
        name: page
        type: integer
      - description: 每页数量，默认10
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '成功响应"  "{\"code\": 0, \"message\": \"success\", \"data\":
            {\"novels\": [...], \"total\": 100, \"page\": 1, \"page_size\": 10}}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: 获取剧本列表
      tags:
      - 内容管理
    post:
      consumes:
      - application/json
      description: 根据资源ID创建剧本，返回剧本ID。这是剧本处理流程的第一步。
      parameters:
      - description: 创建剧本请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/content.CreateNovelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: '成功响应"  "{\"code\": 0, \"message\": \"剧本创建成功\", \"data\": {\"novel_id\":
            \"...\"}}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: 创建剧本
      tags:
      - 内容管理
...
  /api/v1/novels/detail:
    get:
      consumes:
      - application/json
      description: 根据剧本ID获取剧本的详细信息
      parameters:
      - description: 剧本ID
        in: query
        name: novel_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '成功响应"  "{\"code\": 0, \"message\": \"success\", \"data\":
            {\"novel\": {...}}}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: 剧本不存在
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: 获取剧本信息
      tags:
      - 内容管理
  /api/v1/props/images:
    post:
      consumes:
      - application/json
      description: 为小说的所有道具生成图片（抽卡）
      parameters:
      - description: 生成道具图片请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/content.GeneratePropImagesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: 生成道具图片
      tags:
      - 内容管理
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"